TST: split up test_concat.py #37243
 * created test_empty.py

 * changes to test_series.py
